Fix the compile error.

+++ src/common/stats/StatsManager.cpp
@@ -173,7 +173,7 @@ StatusOr<StatsManager::VT> StatsManager::readValue(folly::StringPiece metricName
             return readValue(index, spec.method);
     }
     DLOG(FATAL) << "Unknown method " << static_cast<int>(spec.method);
-    return Status::Error("Unknown method %d.", spec.method);
+    return Status::Error("Unknown method %d.", static_cast<int>(spec.method));
 }
 
 /*static*/ StatusOr<StatsManager::VT> StatsManager::readValue(int32_t index, StatsMethod method) {
